refactor(GAME-26): Replace hook logics
